+++ lib/nadia.ex
@@ -127,6 +127,212 @@ defmodule Nadia do
   def get_me(%Client{} = client), do: api_request(client, "getMe")
 
   @doc """
+  Use this method to log out from the cloud Bot API server before launching the
+  bot locally.
+  Returns `:ok` on success.
+  """
+  @spec log_out() :: :ok | {:error, Error.t()}
+  @spec log_out(Client.t()) :: :ok | {:error, Error.t()}
+  def log_out, do: api_request("logOut")
+  def log_out(%Client{} = client), do: api_request(client, "logOut")
+
+  @doc """
+  Use this method to close the bot instance before moving it from one local
+  server to another.
+  Returns `:ok` on success.
+  """
+  @spec close() :: :ok | {:error, Error.t()}
+  @spec close(Client.t()) :: :ok | {:error, Error.t()}
+  def close, do: api_request("close")
+  def close(%Client{} = client), do: api_request(client, "close")
+
+  @doc """
+  Use this method to change the list of the bot's commands.
+  Returns `:ok` on success.
+  """
+  @spec set_my_commands(list | map | struct | binary) :: :ok | {:error, Error.t()}
+  @spec set_my_commands(list | map | struct | binary, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  @spec set_my_commands(Client.t(), list | map | struct | binary) ::
+          :ok | {:error, Error.t()}
+  @spec set_my_commands(Client.t(), list | map | struct | binary, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def set_my_commands(commands), do: set_my_commands(commands, [])
+
+  def set_my_commands(%Client{} = client, commands) do
+    set_my_commands(client, commands, [])
+  end
+
+  def set_my_commands(commands, options) do
+    api_request(
+      "setMyCommands",
+      request_options(
+        [commands: encode_json_payload(commands)],
+        encode_json_option(options, :scope)
+      )
+    )
+  end
+
+  def set_my_commands(%Client{} = client, commands, options) do
+    api_request(
+      client,
+      "setMyCommands",
+      request_options(
+        [commands: encode_json_payload(commands)],
+        encode_json_option(options, :scope)
+      )
+    )
+  end
+
+  @doc """
+  Use this method to delete the list of the bot's commands for the given scope
+  and user language.
+  Returns `:ok` on success.
+  """
+  @spec delete_my_commands() :: :ok | {:error, Error.t()}
+  @spec delete_my_commands([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec delete_my_commands(Client.t()) :: :ok | {:error, Error.t()}
+  @spec delete_my_commands(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  def delete_my_commands(), do: delete_my_commands([])
+  def delete_my_commands(%Client{} = client), do: delete_my_commands(client, [])
+
+  def delete_my_commands(options),
+    do: api_request("deleteMyCommands", encode_json_option(options, :scope))
+
+  def delete_my_commands(%Client{} = client, options) do
+    api_request(client, "deleteMyCommands", encode_json_option(options, :scope))
+  end
+
+  @doc """
+  Use this method to change the bot's name.
+  Returns `:ok` on success.
+  """
+  @spec set_my_name() :: :ok | {:error, Error.t()}
+  @spec set_my_name([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec set_my_name(Client.t()) :: :ok | {:error, Error.t()}
+  @spec set_my_name(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  def set_my_name(), do: set_my_name([])
+  def set_my_name(%Client{} = client), do: set_my_name(client, [])
+  def set_my_name(options), do: api_request("setMyName", options)
+  def set_my_name(%Client{} = client, options), do: api_request(client, "setMyName", options)
+
+  @doc """
+  Use this method to change the bot's description.
+  Returns `:ok` on success.
+  """
+  @spec set_my_description() :: :ok | {:error, Error.t()}
+  @spec set_my_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec set_my_description(Client.t()) :: :ok | {:error, Error.t()}
+  @spec set_my_description(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  def set_my_description(), do: set_my_description([])
+  def set_my_description(%Client{} = client), do: set_my_description(client, [])
+  def set_my_description(options), do: api_request("setMyDescription", options)
+
+  def set_my_description(%Client{} = client, options) do
+    api_request(client, "setMyDescription", options)
+  end
+
+  @doc """
+  Use this method to change the bot's short description.
+  Returns `:ok` on success.
+  """
+  @spec set_my_short_description() :: :ok | {:error, Error.t()}
+  @spec set_my_short_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec set_my_short_description(Client.t()) :: :ok | {:error, Error.t()}
+  @spec set_my_short_description(Client.t(), [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def set_my_short_description(), do: set_my_short_description([])
+  def set_my_short_description(%Client{} = client), do: set_my_short_description(client, [])
+  def set_my_short_description(options), do: api_request("setMyShortDescription", options)
+
+  def set_my_short_description(%Client{} = client, options) do
+    api_request(client, "setMyShortDescription", options)
+  end
+
+  @doc """
+  Use this method to remove the bot's profile photo.
+  Returns `:ok` on success.
+  """
+  @spec remove_my_profile_photo() :: :ok | {:error, Error.t()}
+  @spec remove_my_profile_photo([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec remove_my_profile_photo(Client.t()) :: :ok | {:error, Error.t()}
+  @spec remove_my_profile_photo(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  def remove_my_profile_photo(), do: remove_my_profile_photo([])
+  def remove_my_profile_photo(%Client{} = client), do: remove_my_profile_photo(client, [])
+  def remove_my_profile_photo(options), do: api_request("removeMyProfilePhoto", options)
+
+  def remove_my_profile_photo(%Client{} = client, options) do
+    api_request(client, "removeMyProfilePhoto", options)
+  end
+
+  @doc """
+  Use this method to change the bot's menu button in a private chat, or the
+  default menu button.
+  Returns `:ok` on success.
+  """
+  @spec set_chat_menu_button() :: :ok | {:error, Error.t()}
+  @spec set_chat_menu_button([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec set_chat_menu_button(Client.t()) :: :ok | {:error, Error.t()}
+  @spec set_chat_menu_button(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  def set_chat_menu_button(), do: set_chat_menu_button([])
+  def set_chat_menu_button(%Client{} = client), do: set_chat_menu_button(client, [])
+
+  def set_chat_menu_button(options) do
+    api_request("setChatMenuButton", encode_json_option(options, :menu_button))
+  end
+
+  def set_chat_menu_button(%Client{} = client, options) do
+    api_request(client, "setChatMenuButton", encode_json_option(options, :menu_button))
+  end
+
+  @doc """
+  Use this method to change the default administrator rights requested by the bot.
+  Returns `:ok` on success.
+  """
+  @spec set_my_default_administrator_rights() :: :ok | {:error, Error.t()}
+  @spec set_my_default_administrator_rights([{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  @spec set_my_default_administrator_rights(Client.t()) :: :ok | {:error, Error.t()}
+  @spec set_my_default_administrator_rights(Client.t(), [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def set_my_default_administrator_rights(), do: set_my_default_administrator_rights([])
+
+  def set_my_default_administrator_rights(%Client{} = client) do
+    set_my_default_administrator_rights(client, [])
+  end
+
+  def set_my_default_administrator_rights(options) do
+    api_request("setMyDefaultAdministratorRights", encode_json_option(options, :rights))
+  end
+
+  def set_my_default_administrator_rights(%Client{} = client, options) do
+    api_request(client, "setMyDefaultAdministratorRights", encode_json_option(options, :rights))
+  end
+
+  @doc """
+  Use this method to change the emoji status for a given user.
+  Returns `:ok` on success.
+  """
+  @spec set_user_emoji_status(integer) :: :ok | {:error, Error.t()}
+  @spec set_user_emoji_status(integer, [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @spec set_user_emoji_status(Client.t(), integer) :: :ok | {:error, Error.t()}
+  @spec set_user_emoji_status(Client.t(), integer, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def set_user_emoji_status(user_id), do: set_user_emoji_status(user_id, [])
+
+  def set_user_emoji_status(%Client{} = client, user_id) do
+    set_user_emoji_status(client, user_id, [])
+  end
+
+  def set_user_emoji_status(user_id, options) do
+    api_request("setUserEmojiStatus", request_options([user_id: user_id], options))
+  end
+
+  def set_user_emoji_status(%Client{} = client, user_id, options) do
+    api_request(client, "setUserEmojiStatus", request_options([user_id: user_id], options))
+  end
+
+  @doc """
   Use this method to send text messages.
   On success, the sent Message is returned.
 
@@ -2892,6 +3098,19 @@ defmodule Nadia do
 
   defp inline_query_result_map(result) do
     for {k, v} <- Map.from_struct(result), v != nil, into: %{}, do: {k, v}
+  end
+
+  defp request_options(required, options) when is_list(options), do: required ++ options
+
+  defp request_options(required, options) when is_map(options),
+    do: Map.merge(Map.new(required), options)
+
+  defp encode_json_option(options, key) when is_list(options) do
+    Keyword.update(options, key, nil, &encode_json_payload/1)
+  end
+
+  defp encode_json_option(options, key) when is_map(options) do
+    Map.update(options, key, nil, &encode_json_payload/1)
   end
 
   defp encode_added_user_ids(options) do