+++ lib/nadia/behaviour.ex
@@ -23,6 +23,56 @@ defmodule Nadia.Behaviour do
 
   @callback get_me :: {:ok, User.t()} | {:error, Error.t()}
   @callback get_me(Client.t()) :: {:ok, User.t()} | {:error, Error.t()}
+  @callback log_out() :: :ok | {:error, Error.t()}
+  @callback log_out(Client.t()) :: :ok | {:error, Error.t()}
+  @callback close() :: :ok | {:error, Error.t()}
+  @callback close(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_my_commands(list | map | struct | binary) :: :ok | {:error, Error.t()}
+  @callback set_my_commands(list | map | struct | binary, [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_my_commands(Client.t(), list | map | struct | binary) ::
+              :ok | {:error, Error.t()}
+  @callback set_my_commands(Client.t(), list | map | struct | binary, [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback delete_my_commands() :: :ok | {:error, Error.t()}
+  @callback delete_my_commands([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback delete_my_commands(Client.t()) :: :ok | {:error, Error.t()}
+  @callback delete_my_commands(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_name() :: :ok | {:error, Error.t()}
+  @callback set_my_name([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_name(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_my_name(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_description() :: :ok | {:error, Error.t()}
+  @callback set_my_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_description(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_my_description(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_short_description() :: :ok | {:error, Error.t()}
+  @callback set_my_short_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_my_short_description(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_my_short_description(Client.t(), [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback remove_my_profile_photo() :: :ok | {:error, Error.t()}
+  @callback remove_my_profile_photo([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback remove_my_profile_photo(Client.t()) :: :ok | {:error, Error.t()}
+  @callback remove_my_profile_photo(Client.t(), [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_chat_menu_button() :: :ok | {:error, Error.t()}
+  @callback set_chat_menu_button([{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback set_chat_menu_button(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_chat_menu_button(Client.t(), [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_my_default_administrator_rights() :: :ok | {:error, Error.t()}
+  @callback set_my_default_administrator_rights([{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_my_default_administrator_rights(Client.t()) :: :ok | {:error, Error.t()}
+  @callback set_my_default_administrator_rights(Client.t(), [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_user_emoji_status(integer) :: :ok | {:error, Error.t()}
+  @callback set_user_emoji_status(integer, [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
+  @callback set_user_emoji_status(Client.t(), integer) :: :ok | {:error, Error.t()}
+  @callback set_user_emoji_status(Client.t(), integer, [{atom, any}] | map) ::
+              :ok | {:error, Error.t()}
   @callback send_message(integer, binary, [{atom, any}]) ::
               {:ok, Message.t()} | {:error, Error.t()}
   @callback send_message(Client.t(), integer | binary, binary, [{atom, any}]) ::
@@ -496,6 +546,37 @@ defmodule Nadia.Behaviour do
               :ok | {:error, Error.t()}
 
   @optional_callbacks get_me: 1,
+                      log_out: 0,
+                      log_out: 1,
+                      close: 0,
+                      close: 1,
+                      set_my_commands: 1,
+                      set_my_commands: 2,
+                      set_my_commands: 3,
+                      delete_my_commands: 0,
+                      delete_my_commands: 1,
+                      delete_my_commands: 2,
+                      set_my_name: 0,
+                      set_my_name: 1,
+                      set_my_name: 2,
+                      set_my_description: 0,
+                      set_my_description: 1,
+                      set_my_description: 2,
+                      set_my_short_description: 0,
+                      set_my_short_description: 1,
+                      set_my_short_description: 2,
+                      remove_my_profile_photo: 0,
+                      remove_my_profile_photo: 1,
+                      remove_my_profile_photo: 2,
+                      set_chat_menu_button: 0,
+                      set_chat_menu_button: 1,
+                      set_chat_menu_button: 2,
+                      set_my_default_administrator_rights: 0,
+                      set_my_default_administrator_rights: 1,
+                      set_my_default_administrator_rights: 2,
+                      set_user_emoji_status: 1,
+                      set_user_emoji_status: 2,
+                      set_user_emoji_status: 3,
                       send_message: 4,
                       forward_message: 4,
                       forward_message: 5,